Remove plural from name
